Made Description a function, not a property.
There's no reason for Description to be loaded
every time a bug is loaded, because it's only
used when viewing the full form of the bug. This
should make it more efficient to LoadBug on any
arbitrary bug and use the helpers, because it now
only needs to parse the directory name.