Simplified GitManager commit handling
There was a lot of hacking around with stash
and reset and dangerous things in the git
commit handler to try and get it to only
commit the issues directory and stash the
index.

It turns out this was unneeded, because
git commit has a "-o" option to only commit
directories passed on the command line and
leave the rest of the index untouched.