Be less aggressive about automatically launching $PAGER
For bug lists of output being sent to the screen (ie. bug list)
it doesn't work very well, since the pager needs to buffer the
output. It also feels weird as a user for reasons I can't explain
to have the result of "bug list" (or "bug purge", or a few others)
automatically paged.