Fixed issue where closing multiple issues closed incorrect issues
Removing the directory was changing the indexes for further lookups
when closing multiple issues, resulting in the wrong bugs being closed.

This makes it a 2 step process (first look up the directories for
everything passed on the command, and then remove them all, instead
of removing them as we find them.), and adds regression tests to make
sure it stays fixed.