Removed bug-serve to move to a different repository
The bug-serve code isn't as mature as the bug code, and
is pretty much a completely independent codebase. It was
never include in any releases unless installed with `go get`.

Being in a separate repository (https://github.com/driusan/bug-serve)
allows development of both bug and bug-serve to move faster.